Reuse functionality in test_autoreduction
